LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/border-image-width-expected.txt


PASS Can set 'border-image-width' to CSS-wide keywords
FAIL Can set 'border-image-width' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'border-image-width' to a length assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a percent assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a number assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to the 'auto' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
PASS Setting 'border-image-width' to a time throws TypeError
PASS Setting 'border-image-width' to an angle throws TypeError
PASS Setting 'border-image-width' to a flexible length throws TypeError
PASS Setting 'border-image-width' to a URL throws TypeError
PASS Setting 'border-image-width' to a transform throws TypeError
PASS 'border-image-width' does not supported '2em 3em'
PASS 'border-image-width' does not supported '5% 15% 10%'
PASS 'border-image-width' does not supported '5% 2em 10% auto'
